FIX: #751, #805, #838, #847, #864, #880, #902, #903, #906, #910, #911, #912, #913, #921 y #925.

+++ sistrahelp/sistrahelp-frontend/src/main/java/es/caib/sistrahelp/frontend/controller/ViewCuadroMando.java
@@ -583,7 +583,7 @@ public class ViewCuadroMando extends ViewControllerBase {
 		this.filtrar();
 	}
 
-	public String calcularEtiquetaTram(Long porcentage, String id) {
+	public String calcularEtiquetaTram(Double porcentage, String id) {
 		String texto = "";
 		if (porcentage > 0) {
 			if (100 - porcentage >= this.umbralAtencionRevisar) {
